init chunk layout generator
