+++ src/bn254/mod.rs
@@ -13,3 +13,5 @@ pub mod ell_coeffs;
 pub mod fp254impl;
 pub mod msm;
 pub mod utils;
+pub mod chunk_primitves;
+pub mod chunk_layout;
